starting scripts
Finalized the GUI and LOG system. Still not complete, but almost finished for the most part. Need GUI for config and about page. Need more display options for LOG system.

Some details on the GUI: Removed the Tab Handler UDF as I found out that you can use the WM NOTIFY and WM COMMAND regularly creating the controls with the UDF.

Added functions and skeletons for Farm Golem. Will work to add previous existing scripts from the old version to this version.

Notes on new functions and files: Added imagesearch files and made the DLL part of the program. It is created on startup to the temporary windows folder.

Some bug fixes on the core functions: The adbCommand was giving a memory leak, but fixed it with proper closing of the process and the standard input output handle.

Some notes: ADB works well, but is slow in detecting changes. Faster alternative is using WINAPI, but the ADB swipe is working well.