bug fixes and navigate quest
Fixed bug where navigating to village hangs and makes the app unresponsive.

Navigate to quest is now coded in the navigate.au3. Also fixed the collectQuest() function to work.

Will be working on Farm Rare and catching system next. Also plan to implement Date Version to specify when the version was updated.